vi: split into implementation files
