+++ src/core/hle/service/vi/manager_display_service.h
@@ -1,0 +1,23 @@
+// SPDX-FileCopyrightText: Copyright 2024 yuzu Emulator Project
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#include "core/hle/service/service.h"
+
+namespace Service::VI {
+
+class IManagerDisplayService final : public ServiceFramework<IManagerDisplayService> {
+public:
+    explicit IManagerDisplayService(Core::System& system_, Nvnflinger::Nvnflinger& nvnflinger_);
+    ~IManagerDisplayService() override;
+
+private:
+    void CloseDisplay(HLERequestContext& ctx);
+    void CreateManagedLayer(HLERequestContext& ctx);
+    void AddToLayerStack(HLERequestContext& ctx);
+    void SetLayerVisibility(HLERequestContext& ctx);
+
+private:
+    Nvnflinger::Nvnflinger& nvnflinger;
+};
+
+} // namespace Service::VI
