+++ src/core/hle/service/vi/manager_root_service.cpp
@@ -1,19 +1,19 @@
 // SPDX-FileCopyrightText: Copyright 2018 yuzu Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "common/logging/log.h"
+#include "core/hle/service/vi/manager_root_service.h"
 #include "core/hle/service/vi/vi.h"
-#include "core/hle/service/vi/vi_m.h"
 #include "core/hle/service/vi/vi_types.h"
 
 namespace Service::VI {
 
-VI_M::VI_M(Core::System& system_, Nvnflinger::Nvnflinger& nv_flinger_,
-           Nvnflinger::HosBinderDriverServer& hos_binder_driver_server_)
+IManagerRootService::IManagerRootService(
+    Core::System& system_, Nvnflinger::Nvnflinger& nv_flinger_,
+    Nvnflinger::HosBinderDriverServer& hos_binder_driver_server_)
     : ServiceFramework{system_, "vi:m"}, nv_flinger{nv_flinger_},
       hos_binder_driver_server{hos_binder_driver_server_} {
     static const FunctionInfo functions[] = {
-        {2, &VI_M::GetDisplayService, "GetDisplayService"},
+        {2, &IManagerRootService::GetDisplayService, "GetDisplayService"},
         {3, nullptr, "GetDisplayServiceWithProxyNameExchange"},
         {100, nullptr, "PrepareFatal"},
         {101, nullptr, "ShowFatal"},
@@ -23,9 +23,9 @@ VI_M::VI_M(Core::System& system_, Nvnflinger::Nvnflinger& nv_flinger_,
     RegisterHandlers(functions);
 }
 
-VI_M::~VI_M() = default;
+IManagerRootService::~IManagerRootService() = default;
 
-void VI_M::GetDisplayService(HLERequestContext& ctx) {
+void IManagerRootService::GetDisplayService(HLERequestContext& ctx) {
     LOG_DEBUG(Service_VI, "called");
 
     detail::GetDisplayServiceImpl(ctx, system, nv_flinger, hos_binder_driver_server,
